pvrvideosink: corruption/jerkiness fixes
1) Use GetDrawableParameters() to get the next client side buffer
(it is double buffered) so we don't immediately try to blit into
the same buffer xorg is reading from.

2) Some related changes to handle invalidated dri2 drawable (such
as window size change or if window is redirected/unredirected by
the window manager).

3) Wait for blits from dst surface before blitting next frame..
this avoids artifacts caused by us writing to a buffer that xorg
is still reading from.  Note that there is a similar fix needed
in the pvr xorg driver to handle the inverse problem.