pvrvideosink: fix incorrect window coordinates
When video is larger than screen resolution, the request to create
window at video size will result in a smaller than requested window.
So get the actual window geometry after the window is created.  This
solves an issue with video appearing "split" with the right half of
the video appearing on the left of the window.