h264parse: disable some optimizations with AVC input
Don't use the new gst_h264_parser_identify_nalu_unchecked and don't split output
AUs when input AUs end (iow when packetized_last=TRUE). This is a workaround for
some interlaced streams that have both fields in the same AVC access unit, but
that apparently need to be converted to one-AU-per-field in byte-stream,au
output.