kmssink: do modeset
Set the mode we actually know what the screen resolution is.  Just
assuming it is the first mode is incorrect and would depend on the
monitor.

We could be a bit more clever and pick the lowest resolution that
is greater than the requested video.  But picking highest resolution
is a safe bet.