pvrvideosink: fix unwanted scaling on cropped videos
When a video with crop rectangle was used, the window was created with
the uncropped size, and kept that way after the crop event was received,
leading to a slight scaling of the video (and likely slight aspect ratio
change as well).

This is fixed by resizing the window to the cropped size when it changes.